Safari-611.1.10/LayoutTests/accessibility/mac/slider-allows-title-ui-element-expected.txt

Color:
This tests that sliders are allowed to have title UI elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS slider.titleUIElement().isEqual(title) is true
PASS title.role is 'AXRole: AXStaticText'
PASS title.childAtIndex(0).role is 'AXRole: AXStaticText'
PASS successfullyParsed is true

TEST COMPLETE
